+++ Tests/Functional/FieldSuggestion/FieldSuggestionWithSeoTest.php
@@ -28,6 +28,7 @@ use TYPO3\CMS\Core\Context\Context;
 use TYPO3\CMS\Core\Database\ConnectionPool;
 use TYPO3\CMS\Core\Http\ServerRequest;
 use TYPO3\CMS\Core\Http\StreamFactory;
+use TYPO3\CMS\Core\Localization\LanguageServiceFactory;
 use TYPO3\CMS\Core\Utility\GeneralUtility;
 use TYPO3\TestingFramework\Core\Functional\FunctionalTestCase;
 
@@ -87,6 +88,7 @@ final class FieldSuggestionWithSeoTest extends FunctionalTestCase
             $rateLimiter,
             GeneralUtility::makeInstance(Context::class),
             new NullLogger(),
+            GeneralUtility::makeInstance(LanguageServiceFactory::class),
         );
         $body     = ['table' => 'pages', 'field' => 'seo_title', 'uid' => 3, 'pid' => 3, 'count' => 3];
         $response = $subject->suggestAction(