+++ Tests/Functional/FieldSuggestion/FieldSuggestionWorkspaceTest.php
@@ -27,6 +27,7 @@ use TYPO3\CMS\Core\Context\Context;
 use TYPO3\CMS\Core\Database\ConnectionPool;
 use TYPO3\CMS\Core\Http\ServerRequest;
 use TYPO3\CMS\Core\Http\StreamFactory;
+use TYPO3\CMS\Core\Localization\LanguageServiceFactory;
 use TYPO3\CMS\Core\Utility\GeneralUtility;
 use TYPO3\TestingFramework\Core\Functional\FunctionalTestCase;
 
@@ -83,6 +84,7 @@ final class FieldSuggestionWorkspaceTest extends FunctionalTestCase
             $rateLimiter,
             GeneralUtility::makeInstance(Context::class),
             new NullLogger(),
+            GeneralUtility::makeInstance(LanguageServiceFactory::class),
         );
         $body = ['table' => 'pages', 'field' => 'description', 'uid' => $uid, 'count' => 3];